requirements.txt
python-docx==0.8.11
markdown==3.1.1
beautifulsoup4==4.9.3
playwright==1.40.0